requirements.txt
langchain_core
langchain_groq
langchain_openai
ofxparse
openai
pandas
plotly
python-dotenv
streamlit
